requirements.txt
accelerate==0.33.0
anthropic==0.25.8
av==13.1.0
beautifulsoup4>=4.12.3
bert-score==0.3.13
bitsandbytes==0.41.3
click==8.1.7
datasets==1.17.0
duckduckgo_search==6.1.0
einops==0.8.0
fastapi[standard]~=0.115.6
google==3.0.0
google-cloud-vision==3.7.4
grpcio==1.60.1
huggingface-hub>=0.23.2
Jinja2~=3.1.3
langdetect>=1.0.9
markdownify~=0.14.1
markdown-pdf
matplotlib>=3.8.2
nltk==3.9.1
numpy>=1.26.0
openai>=1.37.0
orjsonl==1.0.0
pandas>=2.2.0
pillow>=10.3.0
pyparsing~=3.1.1
pytest~=8.3.4
PyYAML>=6.0.1
rank-bm25==0.2.2
requests~=2.32.0
rouge-score==0.1.2
scikit-learn==1.5.1
scipy~=1.12.0
sentence-transformers~=3.4.1
setuptools>=70.0.0
spacy==3.7.4
sty>=1.0.6
tiktoken~=0.7.0
timm==1.0.9
torch~=2.0.1+cu118
torchvision>=0.15.2
transformers>=4.48.0
tqdm>=4.66.3
tweepy~=4.14.0
uvicorn~=0.34.0
wandb==0.18.0
yacs==0.1.8
protobuf~=5.29.3
pydantic~=2.10.6
websockets~=14.2
starlette~=0.45.3
atproto~=0.0.59
prettytable~=3.16.0
aiohttp~=3.11.11